refactor and add
