--- src/main/java/ru/mail/polis/service/kovalkov/Topology.java
@@ -1,14 +1,0 @@
-package ru.mail.polis.service.kovalkov;
-
-import java.nio.ByteBuffer;
-
-public interface Topology<I> {
-    I identifyByKey(ByteBuffer key);
-
-    int nodeCount();
-
-    I[] allNodes();
-
-    boolean isMe(I node);
-
-}
