+++ src/main/java/ru/mail/polis/service/kovalkov/sharding/Topology.java
@@ -1,0 +1,21 @@
+package ru.mail.polis.service.kovalkov.sharding;
+
+import org.jetbrains.annotations.NotNull;
+
+import java.nio.ByteBuffer;
+
+public interface Topology<I> {
+    @NotNull
+    I identifyByKey(@NotNull ByteBuffer key);
+
+    @NotNull
+    I[] replicasFor(@NotNull ByteBuffer key, int replicas);
+
+    int nodeCount();
+
+    @NotNull
+    I[] allNodes();
+
+    boolean isMe(I node);
+
+}
